+++ fungus-cards/src/main/java/com/xq/cards/Constants.java
@@ -1,0 +1,6 @@
+package com.xq.cards;
+
+public class Constants {
+    public static final int NEW_CARD_LIMIT = 1_00_000;
+    public static final String DEFAULT_CARD_TYPE = "CREDIT";
+}