+++ fungus-cards/src/main/java/com/xq/cards/audit/AuditAwareImpl.java
@@ -1,0 +1,14 @@
+package com.xq.cards.audit;
+
+import org.springframework.data.domain.AuditorAware;
+import org.springframework.stereotype.Component;
+
+import java.util.Optional;
+
+@Component("AuditAwareImpl")
+public class AuditAwareImpl implements AuditorAware<String> {
+    @Override
+    public Optional<String> getCurrentAuditor() {
+        return Optional.of("fungus-cards");
+    }
+}