+++ fungus-cards/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.0.0] 
 ### Added
-* Implement card service
+* Implement cards service
 
 ### Changed